// scoutr-vrc/lib/robotevents/robotevents/models/MatchObj.ts
export interface MatchObj {
    /**
     * 
     * @type {number}
     * @memberof MatchObj
     */
    id: number;
    /**
     * 
     * @type {IdInfo}
     * @memberof MatchObj
     */
    event: IdInfo;
    /**
     * 
     * @type {IdInfo}
     * @memberof MatchObj
     */
    division: IdInfo;
    /**
     * 
     * @type {number}
     * @memberof MatchObj
     */
    round: number;
    /**
     * 
     * @type {number}
     * @memberof MatchObj
     */
    instance: number;
    /**
     * 
     * @type {number}
     * @memberof MatchObj
     */
    matchnum: number;
    /**
     * 
     * @type {Date}
     * @memberof MatchObj
     */
    scheduled?: Date;
    /**
     * 
     * @type {Date}
     * @memberof MatchObj
     */
    started?: Date;
    /**
     * 
     * @type {string}
     * @memberof MatchObj
     */
    field?: string;
    /**
     * 
     * @type {boolean}
     * @memberof MatchObj
     */
    scored: boolean;
    /**
     * 
     * @type {string}
     * @memberof MatchObj
     */
    name: string;
    /**
     * 
     * @type {Array<Alliance>}
     * @memberof MatchObj
     */
    alliances: Array<Alliance>;
}