// scoutr-vrc/lib/robotevents/robotevents/models/SkillType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Public Robot Events API
 * An API to access data on Robot Events officially. ## Request Metadata Pagination is performed the same way throughout each pageable endpoint using query parameters `page` and `per_page` (per_page default is 25 with a maximum of 250). <br> Please note that dates should be formatted according to RFC3339. 
 *
 * The version of the OpenAPI document: 1.0.21
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const SkillType = {
    Driver: 'driver',
    Programming: 'programming',
    PackageDeliveryTime: 'package_delivery_time'
} as const;
export type SkillType = typeof SkillType[keyof typeof SkillType];


export function instanceOfSkillType(value: any): boolean {
    for (const key in SkillType) {
        if (Object.prototype.hasOwnProperty.call(SkillType, key)) {
            if (SkillType[key as keyof typeof SkillType] === value) {
                return true;
            }
        }
    }
    return false;
}

export function SkillTypeFromJSON(json: any): SkillType {
    return SkillTypeFromJSONTyped(json, false);
}

export function SkillTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): SkillType {
    return json as SkillType;
}

export function SkillTypeToJSON(value?: SkillType | null): any {
    return value as any;
}

export function SkillTypeToJSONTyped(value: any, ignoreDiscriminator: boolean): SkillType {
    return value as SkillType;
}